pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rl-intraday"
version = "0.1.0"
description = "Production-ready reinforcement learning trading system for intraday futures and options"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "RL Trading Team", email = "trading@example.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "pandas>=2.2.2",
    "numpy>=1.26.4",
    "pyarrow>=16.1.0",
    "polars>=1.4.1",
    "duckdb>=1.1.0",
    "gymnasium>=0.29.1",
    "stable-baselines3>=2.3.2",
    "torch>=2.1",
    "scikit-learn>=1.5.0",
    "scipy>=1.11.0",
    "ib-insync>=0.9.86",
    "databento>=0.36.0",
    "pyyaml>=6.0.2",
    "tqdm>=4.66.4",
    "python-dateutil>=2.9.0",
    "pytz>=2024.1",
    "rich>=13.7.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.6",
    "pytest>=8.3.2",
    "pytest-cov>=5.0.0",
    "black>=24.8.0",
    "ruff>=0.6.8",
    "mypy>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.2",
    "pytest-cov>=5.0.0",
    "black>=24.8.0",
    "ruff>=0.6.8",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]
monitoring = [
    "prometheus-client>=0.20.0",
    "grafana-api>=1.0.3",
]
ta = ["ta-lib>=0.4.28"]

[project.urls]
Homepage = "https://github.com/example/rl-intraday"
Repository = "https://github.com/example/rl-intraday"
Documentation = "https://rl-intraday.readthedocs.io"
"Bug Tracker" = "https://github.com/example/rl-intraday/issues"

[project.scripts]
rl-train = "src.rl.train:main"
rl-paper = "scripts.run_paper_ibkr:main"
rl-pull-data = "scripts.pull_databento_minute:main"

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501", "B008", "C408"]
target-version = "py311"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["B011", "C408"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]